fix: ComixTo - fix getting chapters & pages
Relevant commit : https://github.com/manga-download/haruneko-proprietary/commit/83186c8c6f9cc080838e0ce8e1e1bef59888ab27
