+++ web/src/engine/websites/ComixTo.DRM.d.ts
@@ -1,4 +1,19 @@
+type APIChapter = {
+    id: number;
+    number: number;
+    name: string;
+    group: {
+        name: string;
+    } | null;
+    pages: {
+        baseUrl: string;
+        items: {
+            url: string;
+        }[];
+    };
+};
 export declare class DRMProvider {
-    GetChaptersToken(mangaURL: URL): Promise<string>;
-    GetPagesToken(chapterURL: URL, chapterId: string): Promise<string>;
+    GetChaptersData(mangaURL: URL): Promise<APIChapter[]>;
+    CreatePageLinks(chapterURL: URL): Promise<string[]>;
 }
+export {};
